+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+## 2.3.0 [TODO]
+### Added
+- Get response by-collection-instrument-and-between-datetime endpoint
+### Fixed
+- Get modes by question collectionInstrumentId parameter
+
+## 2.2.3 [2026-02-12]
+### Fixed
+- Fixed INTEGER parsing to use Long instead of int to support large values
+
 ## 2.2.2 [2026-02-05]
 ### Fixed
 - Volumetry don't count responses with collectionInstrumentId